feat: 소요시간을 계산해주는 Calculator 타입 추가 #7
