+++ BankManagerConsoleApp/BankManagerConsoleApp/Employee.swift
@@ -31,26 +31,5 @@ struct Employee {
         
         speaker.speak(when: .finish, number: customer.waitingNumber)
     }
-    
-    func calculate(from openTime: CFAbsoluteTime, to closeTime: CFAbsoluteTime) {
-        let duration = closeTime - openTime
-        
-        guard let workTime = duration.toDecimal else {
-            return
-        }
-        
-        let close = Speaker.Situation.close(time: workTime)
-        speaker.speak(when: close, number: customerCount)
-    }
 }
 
-// MARK : - CFAbsoluteTime Extension
-
-extension CFAbsoluteTime {
-    var toDecimal: String? {
-        let numberFormatter = NumberFormatter()
-        numberFormatter.numberStyle = .decimal
-        numberFormatter.maximumFractionDigits = 2
-        return numberFormatter.string(for: self)
-    }
-}
